pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "empleado-estatal-bot-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.1",
    "praw>=7.7.1",
    "readabilipy>=0.3.0",
    "markdownify>=0.11.6",
    "beautifulsoup4>=4.12.0",
    "colorlog>=6.9.0",
    "curl-cffi>=0.11.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]

[tool.hatch.build.targets.wheel]
packages = ["utils", "infrastructure", "threads"]
